requirements.txt
bambu-connect
obsws_python
python-dotenv